fix(db): avoid invalid truncation errors for DateField

+++ django/db/models/functions/datetime.py
@@ -97,10 +97,10 @@ class Extract(TimezoneMixin, Transform):
                 "TimeField, or DurationField."
             )
         # Passing dates to functions expecting datetimes is most likely a mistake.
-        if isinstance(field, DateField) and copy.lookup_name in (
-            "hour",
-            "minute",
-            "second",
+        if (
+            isinstance(field, DateField)
+            and not isinstance(field, DateTimeField)
+            and copy.lookup_name in ("hour", "minute", "second")
         ):
             raise ValueError(
                 "Cannot extract time component '%s' from DateField '%s'."
@@ -309,9 +309,13 @@ class TruncBase(TimezoneMixin, Transform):
         has_explicit_output_field = (
             class_output_field or field.__class__ is not copy.output_field.__class__
         )
-        if isinstance(field, DateField) and (
-            isinstance(output_field, DateTimeField)
-            or copy.kind in ("hour", "minute", "second", "time")
+        if (
+            isinstance(field, DateField)
+            and not isinstance(field, DateTimeField)
+            and (
+                isinstance(output_field, DateTimeField)
+                or copy.kind in ("hour", "minute", "second", "time")
+            )
         ):
             raise ValueError(
                 "Cannot truncate DateField '%s' to %s."
